Update README.md
A modular big data analysis for the management and development of multi-branch business, which
includes CRM (Customer Relationship Management). The reason behind this project is based on the
recognition of the necessity for this tool, which integrates the needed actions and processes that I
found necessary while working at my current job.